requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==4.2
django-cors-headers==4.9.0
djangorestframework==3.14.0
gunicorn==23.0.0
idna==3.11
mysqlclient==2.2.2
packaging==25.0
pillow==12.0.0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
